+++ web/app/pages/style-guide/module.js
@@ -24,6 +24,7 @@ import WizardPage from './wizard.page';
 import FileUploadDragDropPage from './file-upload-drag-drop.page';
 import TabsPage from './tabs.page';
 import TablesPage from './tables.page';
+import AccordionPage from './accordion.page';
 
 export const routes = [
   <Route
@@ -163,5 +164,11 @@ export const routes = [
     name="style-guide-tables"
     path="/style-guide/tables"
     component={TablesPage}
+  />,
+  <Route
+    key="21"
+    name="style-guide-accordion"
+    path="/style-guide/accordion"
+    component={AccordionPage}
   />
 ];